+++ modules/sd_vae_remote.py
@@ -12,14 +12,25 @@ hf_decode_endpoints = {
     'sd': 'https://q1bj3bpq6kzilnsu.us-east-1.aws.endpoints.huggingface.cloud',
     'sdxl': 'https://x2dmsqunjd6k9prw.us-east-1.aws.endpoints.huggingface.cloud',
     'f1': 'https://whhx50ex1aryqvw6.us-east-1.aws.endpoints.huggingface.cloud',
-    'h1': 'https://whhx50ex1aryqvw6.us-east-1.aws.endpoints.huggingface.cloud',
     'hunyuanvideo': 'https://o7ywnmrahorts457.us-east-1.aws.endpoints.huggingface.cloud',
 }
+hf_decode_endpoints['pixartalpha'] = hf_decode_endpoints['sd']
+hf_decode_endpoints['pixartsigma'] = hf_decode_endpoints['sdxl']
+hf_decode_endpoints['omnigen'] = hf_decode_endpoints['sdxl']
+hf_decode_endpoints['h1'] = hf_decode_endpoints['f1']
+hf_decode_endpoints['lumina2'] = hf_decode_endpoints['f1']
+
 hf_encode_endpoints = {
     'sd': 'https://qc6479g0aac6qwy9.us-east-1.aws.endpoints.huggingface.cloud',
     'sdxl': 'https://xjqqhmyn62rog84g.us-east-1.aws.endpoints.huggingface.cloud',
     'f1': 'https://ptccx55jz97f9zgo.us-east-1.aws.endpoints.huggingface.cloud',
 }
+hf_encode_endpoints['pixartalpha'] = hf_encode_endpoints['sd']
+hf_encode_endpoints['pixartsigma'] = hf_encode_endpoints['sdxl']
+hf_encode_endpoints['omnigen'] = hf_encode_endpoints['sdxl']
+hf_encode_endpoints['h1'] = hf_encode_endpoints['f1']
+hf_encode_endpoints['lumina2'] = hf_encode_endpoints['f1']
+
 dtypes = {
     "float16": torch.float16,
     "float32": torch.float32,
@@ -74,7 +85,7 @@ def remote_decode(latents: torch.Tensor, width: int = 0, height: int = 0, model_
                 params["output_type"] = "pt"
                 params["output_tensor_type"] = "binary"
                 headers["Accept"] = "tensor/binary"
-            if (model_type == 'f1' or model_type == 'h1') and (width > 0) and (height > 0):
+            if model_type in {'f1', 'h1', 'lumina2'} and (width > 0) and (height > 0):
                 params['width'] = width
                 params['height'] = height
             if shared.sd_model.vae is not None and shared.sd_model.vae.config is not None: